refactor: Table 패키지 수정
