+++ src/main/java/persistence/sql/model/Table.java
@@ -1,4 +1,4 @@
-package persistence.sql;
+package persistence.sql.model;
 
 public class Table {
